SparseMatrix: fix index list type in to_mumps_format()

+++ include/finite_element/mumps_solver.hpp
@@ -38,8 +38,8 @@ class MUMPSSolver : public FiniteElement{
 
     private:
     utils::SparseMatrix sK;
-    std::vector<double> rows;
-    std::vector<double> cols;
+    std::vector<int> rows;
+    std::vector<int> cols;
     std::vector<double> vals;
     DMUMPS_STRUC_C config;
